Auth

+++ backend/routes/posts.js
@@ -67,7 +67,8 @@ router.post(
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
 
   post.save()
@@ -97,12 +98,17 @@ router.put(
     _id: req.body.id,
     title: req.body.title,
     content: req.body.content,
-    imagePath
+    imagePath,
+    creator: req.userData.userId
   });
 
-  Post.updateOne({ _id: req.params.id }, post)
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
     .then(result => {
-      res.status(200).json({message: 'Update successful'});
+      if (result.nModified > 0) {
+        res.status(200).json({ message: 'Update successful!' })
+      } else {
+        res.status(401).json({ message: 'Not Authorized!' })
+      }
     })
     .catch(err => console.log(err));
 });
@@ -122,12 +128,13 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id})
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId})
     .then(result => {
-      console.log(result);
-      res.status(200).json({
-        message: 'Post deleted!'
-      });
+      if (result.nModified > 0) {
+        res.status(200).json({ message: 'Deletion successful!' })
+      } else {
+        res.status(401).json({ message: 'Not Authorized!' })
+      }
     });
 });
